separating ComposingContainerDirection in its own file
